+++ tenantforge/migrations/0008_signup_tokens.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS tf_signup_tokens;